Create setTime.js, creating and connecting modules, rename modal.js in app.js deleted unnecessary markup from index.html
